CLI: move most of the functionality into a module
